Updated the slider code with Text View
