+++ app/src/main/java/com/kodeco/android/coordplot/CustomSlider.kt
@@ -1,11 +1,20 @@
 package com.kodeco.android.coordplot
 
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
 import androidx.compose.material.Slider
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import kotlin.math.roundToInt
 
 @Composable
-fun CustomSlider(modifier: Modifier = Modifier, value : Float =0.5f, changeValue : (Float) -> Unit){
+fun CustomSlider(modifier: Modifier = Modifier, value : Float =0.5f, changeValue : (Float) -> Unit,text : String){
+    Text(text = "${text}: ${(value*100).roundToInt()}",
+        Modifier
+            .padding(8.dp)
+            .width(120.dp))
     Slider(
         value = value,
         valueRange = 0.01f..1f,
